move all additional functions from Solver to Algorithms
